🔧 공통으로 사용되는 엔티티들을 패키지로 분류
