+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomPost.java
@@ -1,5 +1,6 @@
 package gaji.service.domain.roomBoard.entity;
 
+import gaji.service.domain.roomBoard.entity.common.*;
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
