+++ src/main/java/gaji/service/domain/roomBoard/entity/common/RoomBoard.java
@@ -1,7 +1,8 @@
-package gaji.service.domain.roomBoard.entity;
+package gaji.service.domain.roomBoard.entity.common;
 
 import gaji.service.domain.enums.RoomPostType;
 import gaji.service.domain.room.entity.Room;
+import gaji.service.domain.roomBoard.entity.RoomPost;
 import jakarta.persistence.*;
 import lombok.*;
