+++ src/main/java/gaji/service/domain/roomBoard/entity/common/RoomComment.java
@@ -1,6 +1,7 @@
-package gaji.service.domain.roomBoard.entity;
+package gaji.service.domain.roomBoard.entity.common;
 
 import gaji.service.domain.enums.UserAlarmTypeEnum;
+import gaji.service.domain.roomBoard.entity.RoomPost;
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
