+++ src/main/java/gaji/service/domain/roomBoard/entity/common/RoomPostFile.java
@@ -1,5 +1,6 @@
-package gaji.service.domain.roomBoard.entity;
+package gaji.service.domain.roomBoard.entity.common;
 
+import gaji.service.domain.roomBoard.entity.RoomPost;
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -9,16 +10,19 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RoomCommentLikes {
+public class RoomPostFile {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id")
-    private RoomComment roomComment;
+    @JoinColumn(name = "post_id")
+    private RoomPost roomPost;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
+
+    private String path;
+
 }
