+++ src/main/java/gaji/service/domain/roomBoard/service/RoomPostCommandServiceImpl.java
@@ -3,7 +3,7 @@ package gaji.service.domain.roomBoard.service;
 import gaji.service.domain.room.entity.Room;
 import gaji.service.domain.room.service.RoomQueryService;
 import gaji.service.domain.roomBoard.converter.RoomPostConverter;
-import gaji.service.domain.roomBoard.entity.RoomBoard;
+import gaji.service.domain.roomBoard.entity.common.RoomBoard;
 import gaji.service.domain.roomBoard.entity.RoomPost;
 import gaji.service.domain.roomBoard.repository.RoomBoardRepository;
 import gaji.service.domain.roomBoard.repository.RoomPostRepository;
