+++ src/main/java/gaji/service/domain/user/entity/User.java
@@ -14,6 +14,7 @@ import gaji.service.domain.room.entity.Room;
 import gaji.service.domain.room.entity.RoomEvent;
 import gaji.service.domain.room.entity.VoiceChatUser;
 import gaji.service.domain.roomBoard.entity.*;
+import gaji.service.domain.roomBoard.entity.common.*;
 import gaji.service.domain.studyMate.entity.*;
 import gaji.service.oauth2.dto.TransferUserDTO;
 import jakarta.persistence.*;
